organize event listener setup into seperate directory; implement todo delete button
