+++ src/vanilla/scripts/app.ts
@@ -1,6 +1,6 @@
 import "../../common/styles/styles.scss"
 
-import { attachEventListeners } from "./eventListeners"
+import { attachEventListeners } from "./events"
 import { loadState, saveState } from "./george"
 import { render } from "./render"
 import { createStore } from "./store"
@@ -8,12 +8,6 @@ import { createStore } from "./store"
 const { dispatch, subscribe, getState } = createStore(loadState())
 
 const saveSubscription = subscribe(saveState)
-
-// const spySubscription = subscribe((state) => {
-//   const spy = <HTMLElement>document.getElementById("spy")
-
-//   spy.innerHTML = JSON.stringify(state, null, 2)
-// })
 
 const renderSubscription = subscribe(render)
